teachme-live-claude/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.0
python-multipart==0.0.9
pydantic==2.9.2
pydantic-settings==2.6.1
python-dotenv==1.0.1
anthropic>=0.39.0
google-cloud-speech>=2.28.0
google-cloud-texttospeech>=2.16.0
websockets>=12.0
aiohttp>=3.9.0
numpy>=1.24.0
pydub>=0.25.1
sympy>=1.13.3